importHtml/Text: minor refactor
